Update index.html
updated index adding all the code which has been written to design this specific home page of the website. coding for buttons and navigation to work still needs to be done, however this has been uploaded to be an outline to my group partners to add onto for their respective sections of the website.